+++ src/features/task-form/components/TaskForm/index.tsx
@@ -6,24 +6,9 @@ import {
   useId,
   useState,
 } from "react";
+import type { TaskFormValues } from "@/features/task-form/types";
 import type { Column, Priority, Task } from "@/types/task";
 import { ParentTaskSelect } from "../ParentTaskSelect";
-
-/** TaskForm から送信される値 */
-export type TaskFormValues = {
-  /** タイトル（必須、空文字不可） */
-  title: string;
-  /** ステータス（必須） */
-  status: string;
-  /** 優先度（任意） */
-  priority?: Priority;
-  /** ラベル一覧 */
-  labels: string[];
-  /** 親タスクのファイルパス（任意） */
-  parent?: string;
-  /** 本文（Markdown） */
-  body: string;
-};
 
 type TaskFormProps = {
   /** 選択肢となるカラム一覧 */
