+++ src/features/task-form/index.ts
@@ -1,2 +1,2 @@
 export { TaskCreateModal } from "./components/TaskCreateModal";
-export type { TaskFormValues } from "./components/TaskForm";
+export type { TaskFormValues } from "./types";
